requirements.txt
Babel==2.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
flask-babel==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
Werkzeug==3.0.1
